Make cancel request keys longer, bump minor protocol version
Currently, cancel request key is 32-bit token, which isn't very much
entropy. If you want to cancel another session's query, you can
brute-force it. In most environments, an unauthorized cancellation of
a query isn't very serious, but it nevertheless would be nice to have
more protection from it.

To make it harder to guess the key, make it larger.

The new longer key length is not hardcoded in the protocol anymore,
the client is expected to deal with variable length keys, up to some
reasonable upper limit (TODO: document the maximum). This flexibility
allows e.g. a connection pooler to add more information to the cancel
key, which might be useful for finding the connection.

This bumps the protocol version from 3.0 to 3.1. The server and libpq
still both understand version 3.0, for interoperability with old
client and server versions. If an old client connects to the server,
the server falls back to generate a short 32-bit token like before.

Note that the minor protocol version negotiation was added to the
server in version 9.3. Connecting to PostgreSQL 9.2 with new libpq
will fail with "unsupported frontend protocol" error. If we wanted to
stay compatible, we could teach libpq to reconnect with protocol
version 3.0, but I don't think that's necessary.

Reviewed-by: Jelte Fennema-Nio
Discussion: https://www.postgresql.org/message-id/508d0505-8b7a-4864-a681-e7e5edfe32aa@iki.fi