Speed up TransactionIdIsCurrentTransactionId() with lots of subxacts
TransactionIdIsCurrentTransactionId() would traverse through the whole
stack of TransactionStates, making it O(n) where n is the number of
in-progress subtransactions. That can be pretty slow.

Replace the childXids array on each nesting level with one big array
(CurrentTransactionXids), that holds the XIDs of all
subtransactions. This way, TransactionidIsCurrentTransactionId() can
perform one binary search over all the XIDs, regardless of how many
subtransations are active. For each nesting level, store an offset
into that big array (totalXids), indicating which portion of the array
"belongs" to that level. When a subtransaction commits, instead of
copying all the child XIDs to the parent's array, just update the
parent's totalXids to include the children.