Rename some signal and interrupt handling functions for consistency
The usual pattern for handling a signal is that the signal handler
sets a flag and calls SetLatch(MyLatch) to wake up the process, and
CHECK_FOR_INTERRUPTS() or other code that is part of a wait loop calls
another function to deal with it. The naming of the functions involved
was a bit inconsistent, however. CHECK_FOR_INTERRUPTS() calls
ProcessInterrupts() to do the heavy-lifting, but the analogous
functions in aux processes were called HandleMainLoopInterrupts(),
HandleStartupProcInterrupts(), etc. Similarly, most subroutines of
ProcessInterrupts() were called Process*(), but some were called
Handle*().

To make things less confusing, rename all the functions that are part
of the overall signal/interrupt handling system but are not executed
in a signal handler to e.g. ProcessSomething(), rather than
HandleSomething().