Replace ProcSignals and other ad hoc signals with interrupts
This replaces the INTERRUPT_GENERAL, or setting the process's latch
before that, with more fine-grained interrupts. All the ProcSignals
are converted into interrupts, as well as things like timers and
config-reload requests.

Most callers of WaitInterrupt now use INTERRUPT_CFI_MASK, which is a
bitmask that includes all the interrupts that are handled by
CHECK_FOR_INTERRUPTS(). CHECK_FOR_INTERRUPTS() now clears the
interrupts bits that it processes, the caller is no longer required to
do ClearInterrupt (or ResetLatch) for those bits. If you wake up for
other interrupts that are not handled by CHECK_FOR_INTERRUPTS(), you
still need to clear those.

Now that there are separate bits for different things, it's possible
to do e.g. WaitInterrupt(INTERRUPT_CONFIG_RELOAD |
INTERRUPT_QUERY_CANCEL) to wait just for a config reload request or
query cancellation, and not wake up on other events like a shutdown
request. That's not very useful in practice, as CHECK_FOR_INTERRUPTS()
still processes those interrupts if you call it, and you would want to
process all the standard interrupts promptly anyway. It might be
useful to check for specific interrupts with InterruptPending(),
without immediately processing them, however. spgdoinsert() does
something like that. In this commit I didn't change its overall logic,
but it probably could be made more fine-grained now.

More importantly, you can now have additional interrupts that are
processed less frequently and do not wake up the backend when it's not
prepared to process them. For example, the SINVAL_CATCHUP interrupt is
only processed when the backend is idle. Previously, a catchup request
would interrupt the process when it was not idle but was waiting on
other things; now it only wakes up when the backend is idle and is
ready to process it. That's not too significant for catchup interrupts
because they don't occur that often anyway for a few extra wakeups to
matter, but it's still nice and could be more significant with other
interrupts.

There's still a general-purpose INTERRUPT_WAIT_WAKEUP interrupt that
is multiplexed for many different purposes in different processes, for
example to wake up the walwriter when it has work to do, or to wake up
processes waiting on a condition variable. The common property of
those uses is that there's some other flag or condition that is
checked on each wakeup, the wakeup interrupt itself means merely that
something interesting might've happened.

TODO:

 * ShutdownRequestPending became INTERRUPT_SHUTDOWN_AUX and
   ProcDiePending became INTERRUPT_DIE, but do we really need two
   different interrupts for requesting a process to terminate?

 * There are only a few interrupts bits left. Consolidate the existing
   ones, switch to a 64-bit atomic, or mulitplex some of them?

XXX: original text from Thomas's patch:

"ProcSignals" were a two-step way for one backend to ask another backend
to do something, by multiplexing SIGUSR1.  Historically, real work was
done in the SIGUSR1 handler, but commit 0da096d7 removed the last case
of that.  Now the handler just set "interrupt pending" flags, called
SetLatch(MyLatch) to close races, and then did all the interesting work
at the next CHECK_FOR_INTERRUPTS() call.

This commit removes the ProcSignal step, replacing ProcSignals with
Interrupts.

Benefits of this refactoring include removing unnecessary concepts and
terminology, avoiding the risk of people putting buggy code back into
signal handlers, removing a whole lot of global variables, removing
dependencies on PIDs and the process model, removing risks of sending
a signal that has default action of termination to the wrong process,
and taking a step towards being able remove the fake signal system
from the Windows port.

Notable changes:

 * all calls to SendProcSignal(pid, PROCSIG_XXX) become
   SendInterrupt(INTERRUPT_XXX, procno)
 * all XxxPending global variables are gone; they are represented
   as bits in MyProc->pending_interrupts
 * the SIGUSR1 handler is gone
 * places that know the PIDs of other backends must now use procnos
   instead

The only code left in src/backend/storage/ipc/procsignal.c relates to
ProcSignalBarriers.  XXX Those should perhaps be renamed.

SIGUSR1 can now be set to SIG_IGN in most backends.  Some special cases
remain, that were not using the general sigusr1_handler (now removed).

Author: Thomas Munro <thomas.munro@gmail.com>
Author: Heikki Linnakangas <heikki.linnakangas@iki.fi>
Discussion: https://postgr.es/m/CA%2BhUKG%2B3MkS21yK4jL4cgZywdnnGKiBg0jatoV6kzaniBmcqbQ%40mail.gmail.com