Bump protocol version to 3.2
In preparation of new additions to the protocol in the next commit,
this bumps the minor protocol version number. Until then, the old 3.0
and the 3.2 versions are still identical.

The server and libpq are both still compatible with protocol version
3.0, and libpq still defaults to version 3.0. You have to explicitly
ask for the new protocol by setting max_protocol_version=latest or
min_protocol_version=latest, the difference between the two being that
with min_protocol_version=latest, libpq will reject the connection
when connecting to an older server that doesn't support the new
protocol, while with max_protocol_version=latest, it will downgrade to
the old protocol.

Instead of bumping the version number to 3.1, which would be the next
minor version, we skip that one and bump straight to 3.2. The reason
for this is that many PgBouncer releases have, due to an off-by-one
bug, reported 3.1 as supported. These versions would interpret 3.1 as
equivalent to 3.0. So if we would now add extra messages to the 3.1
protocol, clients would succeed to connect to PgBouncer, but would
then cause connection failures when sending such new messages.

So instead of bumping to 3.1, we bump the protocol version to 3.2, for
which these buggy PgBouncer releases will correctly close the connection
at the startup packet. It's a bit strange to skip a version number due
to a bug in a third-party application, but PgBouncer is used widely
enough that it seems worth it to not cause user confusion when
connecting to recent versions of it. Especially since skipping a single
minor version number in the protocol versioning doesn't really cost us
anything. So, while this is not the most theoretically sound decission,
it is the most pragmatic one.

Author: Jelte Fennema-Nio <postgres@jeltef.nl>
Discussion: https://www.postgresql.org/message-id/CAGECzQTfc_O%2BHXqAo5_-xG4r3EFVsTefUeQzSvhEyyLDba-O9w%40mail.gmail.com
Discussion: https://www.postgresql.org/message-id/CAGECzQRbAGqJnnJJxTdKewTsNOovUt4bsx3NFfofz3m2j-t7tA@mail.gmail.com
Discussion: https://github.com/pgbouncer/pgbouncer/pull/1007