Invent custom pg_getopt_ctx that is thread-safe
The standard getopt(3) function is not re-entrant nor
thread-safe. That's OK for current usage, but it's one more little
thing we need to change in order to make the server multi-threaded.

There's no standard getopt_r() function on any platform, because
command line arguments are usually parsed early when you start a
program, usually before launching any threads, so there's not much
need for it. However, we call it at backend startup to parse options
from the startup packet. We are therefore free to define our own.

The pg_getopt_start/next() implementation is based on the old getopt
implementation, I just gathered all the state variables to a struct.
The non-re-entrant getopt() function is now a wrapper around the
custom re-entrant variant, on platforms that don't have getopt(3).
getopt_long() is not used in the server, so we don't need to provide a
re-entrant variant of that.