Use strtoi64() in pgbench, replacing its open-coded implementation
Makes the code a little simpler.

The old implementation accepted trailing whitespace, but that seemed
unnecessary. Firstly, its sibling function for parsing decimals,
strtodouble(), does not accept trailing whitespace. Secondly, none of
the callers can pass a string with trailing whitespace to it.

In the passing, check specifically for ERANGE before printing the "out
of range" error. On some systems, strtoul() and strtod() return INVAL
on an empty string, and "invalid input syntax" is more appropriate for
that than "out of range". For the existing strtodouble() function this
is purely academical because it's never actually called with
errorOk==false, but let's be tidy. (Perhaps we should remove the dead
codepaths altogether, but I'll leave that for another day.)

Reviewed-by: Chao Li <li.evan.chao@gmail.com>
Reviewed-by: Yuefei Shi <shiyuefei1004@gmail.com>
Reviewed-by: Neil Chen <carpenter.nail.cz@gmail.com>
Discussion: https://www.postgresql.org/message-id/861dd5bd-f2c9-4ff5-8aa0-f82bdb75ec1f@iki.fi